Move character-related structs into their own module
