+++ src/lib.rs
@@ -4,14 +4,12 @@ pub mod commands;
 mod connection;
 mod listener;
 mod object;
-mod player;
-mod pronoun;
 mod room;
 pub mod util;
 pub mod world;
 
 pub use area::Area;
-pub use character::{CharId, Character};
+pub use character::{CharId, Character, PlayerRecord};
 pub use commands::lookup_command;
 pub use connection::{Connection, ConnectionBuilder};
 pub use listener::listen;
@@ -19,7 +17,5 @@ pub use object::{
     AllObjectsAdapter, Object, ObjectDef, ObjectId, ObjectInRoomAdapter, ObjectOnCharAdapter,
     ObjectType,
 };
-pub use player::{Player, PlayerRecord};
-pub use pronoun::Pronoun;
 pub use room::{Exit, Room, RoomId};
 pub use world::World;
